test: run daemon /exec result tests against the real driver in lab-runner
Four homeboy-core daemon tests assert on the *result* of actually running the
requested command through the daemon `/exec` endpoint — succeeded/failed status,
captured stdout/stderr and exit code, and the captured-patch delta artifact. The
`/exec` endpoint drives execution through the `RunnerExecDriver` hook (#8632),
whose real implementation was extracted into the runner crate (#8698), so in
homeboy-core these tests hit the fail-closed 400 with no driver — and no core-only
stub can produce a real command result without becoming a test double.

Move them to `homeboy-lab-runner`, which registers the real driver
(`register_runner_daemon_exec_driver`), so the command runs for real and the
result events, exit codes, stdout/stderr, and patch artifacts are exercised end
to end:

- `daemon_exec_does_not_require_runner_config_on_daemon_host`
- `exec_applies_request_env_to_daemon_command`
- `exec_failed_command_marks_job_failed_after_result_event`
- `exec_capture_patch_records_remote_delta_artifact`

To keep the in-process dispatch the core tests used (no HTTP listener needed),
expose `homeboy_core::daemon::route_with_body` — the same trusted-local router the
authenticated network entry point uses. All four tests pass in lab-runner with
the real driver, and the remaining core daemon `/exec` tests (which only assert
enqueue behavior) are unaffected.
