+++ crates/homeboy-lab-runner/src/execution/tests/mod.rs
@@ -8,6 +8,7 @@ use homeboy_core::defaults::AgentTaskSecretSource;
 use homeboy_core::error::ErrorCode;
 use homeboy_core::server::{self, RunnerPolicy, RunnerSecretEnvRef, RunnerSettings};
 
+mod daemon_exec;
 mod exec;
 mod handoff;
 mod policy;
